etl/requirements.txt
backoff==2.2.1
elasticsearch==8.5.3
psycopg2_binary==2.9
python-dotenv==0.21.0
redis==4.4.0
pydantic==1.10.4
orjson==3.8.5
